This is a sample classifier hierarchy w/ fixtures
The classes_dump.json can be loaded into a clean/running Classifier
instance through the Hierarchy Import.

Each of the files in the fixtures/ directory contains facts
for the node whose hostname is the same as the file's name;
these can be used by POST'ing them to the Classification
Explanation endpoint.